requirements.txt
certifi==2019.11.28
chardet==3.0.4
dataclasses==0.7
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.37.1
django-heroku==0.3.1
django-leaflet==0.26.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.9
oauthlib==3.1.0
Pillow==7.0.0
psycopg2==2.8.4
pyproj==2.5.0
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.2.4
urllib3==1.25.8
whitenoise==5.0.1
